refactor(operators): migrate to NumPy-aligned object pattern
Operator Overload Migration:
- Arithmetic (+, -, *, /, %): Replace 120 explicit scalar overloads with
  10 object-based overloads using np.asanyarray() for type conversion
- Bitwise (&, |, ^): Already using object pattern, verified working
- Comparison (==, !=, <, <=, >, >=): Already using object pattern
- Unary (-, +, ~, !): Implemented, no object pattern needed

Type Conversion Alignment (NumPy behavior):
- scalar → NDArray: implicit (safe, creates 0-d array)
- NDArray → scalar: explicit (requires 0-d, throws IncorrectShapeException)
- This matches NumPy's int(arr), float(arr), bool(arr) pattern

Null Check Migration:
- Changed == null to is null throughout codebase
- Reason: == now returns NDArray<bool> for element-wise comparison
- is null performs identity check (matches NumPy's "is None" pattern)

Code Reduction:
- NDArray.Primitive.cs: 159 → 42 lines (74% reduction)
- Total: ~150 explicit overloads → ~40 object-based overloads

Added:
- NDArray.BitwiseNot.cs: ~ operator implementation
- NDArray.XOR.cs: ^ operator with object pattern
- docs/OPERATOR_ALIGNMENT.md: Comprehensive operator documentation

Fixed implicit conversions:
- Added missing implicit operator NDArray(byte)
- Changed ushort from explicit to implicit

+++ src/NumSharp.Core/Sorting_Searching_Counting/np.argmax.cs
@@ -20,7 +20,7 @@
         /// <returns>Index of the maximum value in the flattened array.</returns>
         /// <remarks>https://numpy.org/doc/stable/reference/generated/numpy.argmax.html</remarks>
         public static long argmax(NDArray a)
-            => a.TensorEngine.ArgMax(a);
+            => (long)a.TensorEngine.ArgMax(a);
 
         /// <summary>
         ///     Returns the indices of the minimum values along an axis.
@@ -40,6 +40,6 @@
         /// <returns>Index of the minimum value in the flattened array.</returns>
         /// <remarks>https://numpy.org/doc/stable/reference/generated/numpy.argmin.html</remarks>
         public static long argmin(NDArray a)
-            => a.TensorEngine.ArgMin(a);
+            => (long)a.TensorEngine.ArgMin(a);
     }
 }
